Move contained flow constants from common to own file
