+++ client/src/constants/common.ts
@@ -51,22 +51,6 @@ export const MANAGE_PAGE_REDIRECT = {
   domain: ROUTES.FIO_DOMAINS,
 };
 
-export const CONTAINED_FLOW_ACTIONS = {
-  SIGNNFT: CONFIRM_PIN_ACTIONS.SIGN_NFT,
-};
-
-export const CONTAINED_FLOW_ACTIONS_TO_ROUTES = {
-  [CONTAINED_FLOW_ACTIONS.SIGNNFT]: ROUTES.CONTAINED_FLOW_SIGN_NFT,
-};
-
-export const CONTAINED_FLOW_STEPS = {
-  INIT: 'INIT',
-  CREATE_ACCOUNT: 'CREATE_ACCOUNT',
-  REGISTRATION: 'REGISTRATION',
-  ACTION: 'ACTION',
-  FINISH: 'FINISH',
-};
-
 export const NFT_TOKEN_ITEM_PROPS_ORDER: NFTTokenItemProps[] = [
   'fioAddress',
   'chainCode',
